requirements.txt
paho-mqtt
pyserial
pymysql

